refactor db-user into its own class
